Revert "Pull request for integration of build tasks"
