+++ README.md
@@ -1,2 +1,2 @@
 # Shopping List App
-This is a shopping list app that allows for the creation of various lists with their own items. It uses Html, CSS, Jquery, Javascript, LocalStorage, Chokidar, Less, Clean CSS, Webpack, and Babel. You can see how it works on https://jayzamazing.github.io/Shopping-List-App.
+This is a shopping list app that allows for the creation of various lists with their own items. It uses Html, Css, Jquery, Javascript, and LocalStorage. You can see how it works on https://jayzamazing.github.io/Shopping-List-App.
