+++ css/style.css
@@ -1,0 +1,127 @@
+html, body {
+  -webkit-box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  box-sizing: border-box;
+  margin:0;
+  padding:0;
+  color: #261a0d;
+}
+body {
+  background-image: url('../img/shopping-center-iot.jpg');
+  -webkit-background-size: cover;
+  -moz-background-size: cover;
+  -o-background-size: cover;
+  background-size: cover;
+  height: 45.938rem;
+  background-position: center bottom;
+  overflow: auto;
+  padding:0;
+  margin: auto;
+  font-family: 'Architects Daughter', cursive;
+  background-repeat: no-repeat;
+}
+h1, h3 {
+  font-family: 'Bangers', cursive;
+}
+h1 {
+  font-size: 64px;
+}
+*, *:before, *:after {
+  -webkit-box-sizing: inherit;
+  -moz-box-sizing: inherit;
+  box-sizing: inherit;
+}
+.content {
+  margin: 0 auto;
+  padding: 1rem;
+
+}
+.content-background {
+  background-color: #dfbf9f;
+  border-bottom-right-radius: 25px;
+  border-top-left-radius: 25px;
+  border: 1px solid black;
+  opacity: 0.9;
+  filter: alpha(opacity=60);
+}
+.content ul {
+  padding: 0;
+}
+.content li {
+  display: block;
+  margin-bottom: 5px;
+}
+.content i {
+  border-radius: 1.25rem;
+  width: 1rem;
+  height: 1rem;
+  box-shadow: 0px 0px 3px #18191b;
+  padding-left: .1rem;
+}
+.listNames, .listItems {
+  float: left;
+}
+.listNames .error, .listItems .error {
+  border-color: red;
+}
+.title, .listItemsTitle {
+  text-align: center;
+}
+form {
+  text-align: center;
+}
+.fa-times {
+  color: red;
+  float: right;
+}
+.highlight {
+  background-color: #cb9b67;
+}
+.storageErrorStatus {
+  visibility: hidden;
+  color: red;
+}
+.hidden {
+  display: none !important;
+}
+@media only screen and (max-width: 48rem) {
+  section, aside {
+    width: 100%;
+  }
+  aside .content {
+    margin-bottom: 1rem;
+  }
+  h1 {
+    font-size: 32px;
+  }
+}
+@media only screen and (min-width: 48rem) {
+  body {
+    width: 45rem;
+  }
+  aside .content {
+      width: 12rem;
+      margin-right: 3rem;
+  }
+  section .content {
+    width: 30rem;
+  }
+  input.listName {
+    width: 5rem;
+  }
+}
+@media only screen and (min-width: 75rem) {
+  body {
+    width: 60rem;
+  }
+  aside .content {
+    width: 15rem;
+    margin-right: 5rem;
+  }
+  section .content {
+    width: 40rem;
+  }
+  input.listName {
+    width: 7rem;
+  }
+}
