+++ index.html
@@ -4,14 +4,14 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <title>Awesome Shopping List</title>
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/4.1.1/normalize.css" type="text/css">
+  <link rel="stylesheet" href="node_modules/normalize.css/normalize.css" type="text/css">
   <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css" rel="stylesheet">
   <link href='https://fonts.googleapis.com/css?family=Bangers|Architects+Daughter' rel='stylesheet' type='text/css'>
-  <link rel="stylesheet" href="build/shopping-list-app.1.0.0.min.css" type="text/css">
+  <link rel="stylesheet" href="css/style.css" type="text/css">
   <link rel="shortcut icon" type="image/x-icon" href="favicon.ico?v=1" />
 </head>
 <body>
-  <header class="title center-text">
+  <header class="title">
     <div class="content">
       <h1 class="titleImg">Awesome Shopping List</h1>
       <h3 class="storageErrorStatus">Local Storage Not Available</h3>
@@ -21,7 +21,7 @@
     <div class="content content-background">
       <h3 class="listTitle">Shopping Lists</h3>
       <hr>
-      <form action="" id="listNameForm center-text">
+      <form action="" id="listNameForm">
         <input type="text" id="listName" class="listName" maxlength="12">
         <input type="submit" value="Submit" id="listNameButton">
       </form>
@@ -31,10 +31,10 @@
   </aside>
   <section class="listItems">
     <div class="content content-background">
-      <h3 class="listItemsTitle center-text">Shopping List Items</h3>
+      <h3 class="listItemsTitle">Shopping List Items</h3>
       <hr>
       <div class="addListItem">
-        <form action="" id="itemNameForm center-text">
+        <form action="" id="itemNameForm">
           <input type="text" class="itemName" id="itemName" maxlength="45">
           <input type="submit" value="Submit" id="itemNameButton">
         </form>
@@ -43,6 +43,7 @@
       </ul>
     </div>
   </section>
-	<script src="build/shopping-list-app.1.0.0.js"></script>
+  <script src="https://code.jquery.com/jquery.min.js"></script>
+	<script src="js/app.js"></script>
 </body>
 </html>
