--- package.json
@@ -1,63 +1,0 @@
-{
-  "name": "shopping-list-app",
-  "version": "1.0.0",
-  "description": "This is a shopping list app that allows for the creation of various lists with their own items. It uses Html, Css, Jquery, Javascript, and LocalStorage.",
-  "main": "js/index.js",
-  "dependencies": {
-    "jquery": "3.0.0",
-    "normalize.css": "4.1.1"
-  },
-  "devDependencies": {
-    "babel-core": "^6.10.4",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.9.0",
-    "chokidar": "1.6.0",
-    "less": "2.7.1",
-    "less-plugin-clean-css": "1.5.1",
-    "webpack": "1.13.1"
-  },
-  "scripts": {
-    "mkdir": "mkdir -p build",
-    "build": "npm run clean && npm run mkdir && npm run build:html && npm run build:css && npm run build:js",
-    "watch": "npm run watch:html & npm run watch:css & npm run watch:js",
-    "clean": "rm -rf build",
-    "build:html": "npm run clean:html && cp index.html build/",
-    "watch:html": "npm run build:html && chokidar index.html -c 'npm run build:html'",
-    "clean:html": "rm -f build/index.html",
-    "build:css": "npm run build:css:unminified && npm run build:css:minified",
-    "watch:css": "npm run watch:css:unminified & npm run watch:css:minified",
-    "clean:css": "npm run clean:css:unminified && npm run clean:css:minified",
-    "build:css:unminified": "npm run clean:css:unminified && lessc --source-map css/index.less build/$npm_package_name.$npm_package_version.css",
-    "watch:css:unminified": "npm run build:css:unminified && chokidar 'css/**/*.(css|less)' -c 'npm run build:css:unminified'",
-    "clean:css:unminified": "rm -f build/$npm_package_name.$npm_package_version.css build/$npm_package_name.$npm_package_version.css.map",
-    "build:css:minified": "npm run clean:css:minified && lessc --clean-css --source-map css/index.less build/$npm_package_name.$npm_package_version.min.css",
-    "watch:css:minified": "npm run build:css:minified && chokidar 'css/**/*.(css|less)' -c 'npm run build:css:minified'",
-    "clean:css:minified": "rm -f build/$npm_package_name.$npm_package_version.min.css build/$npm_package_name.$npm_package_version.min.css.map",
-    "build:js": "npm run clean:js && webpack",
-    "watch:js": "npm run build:js && webpack --watch",
-    "clean:js": "rm -f build/$npm_package_name.$npm_package_version.js build/$npm_package_name.$npm_package_version.js.map"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jayzamazing/Shopping-List-App.git"
-  },
-  "keywords": [
-    "Html",
-    "Css",
-    "Jquery",
-    "Javascript",
-    "LocalStorage",
-    "Normalize",
-    "Chokidar",
-    "Less",
-    "Clean CSS",
-    "Webpack",
-    "Babel"
-  ],
-  "author": "Adrian J Lopez <adrianjaylopez@gmail.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/jayzamazing/Shopping-List-App/issues"
-  },
-  "homepage": "https://github.com/jayzamazing/Shopping-List-App#readme"
-}
